--- custom-plugin-decorators/spread-security-to-operations/plugin.js
@@ -1,12 +1,0 @@
-import spreadSecurityToOperations from "./decorator";
-
-export default function plugin() {
-  return {
-    id: "security-plugin",
-    decorators: {
-      oas3: {
-        "spread-security-to-operations": spreadSecurityToOperations,
-      },
-    },
-  };
-}
